% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tictok.R
\name{tiktok}
\alias{tiktok}
\alias{tiktok_embed}
\alias{tiktok_md}
\alias{tiktok_html}
\title{Embed a TikTok Video}
\usage{
tiktok_embed(url)

tiktok_md(tt, ...)

tiktok_html(tt, include_player = TRUE, ...)
}
\arguments{
\item{url}{A TikTok video URL}

\item{tt}{A TikTok video URL or \code{tiktok_embed()} result}

\item{...}{Ignored}

\item{include_player}{Include the TikTok embed JavaScript that builds the
rich HTML player (and probably adds a considerable amount of user tracking)}
}
\value{
\code{tiktok_embed}: Returns the TikTok oEmbed API response. The returned
object has a \code{knit_print()} method for automatic embedding in R Markdown.

\code{tiktok_md}: A Markdown representation of the TikTok video, in the
format \verb{![<title> by <author>](<thumbnail_image_url>)}

\code{tiktok_html}: The TikTok embedding HTML as an \code{htmltools::tagList}
}
\description{
Embeds a TikTok video in an R Markdown document. In rich HTML formats, the
full TikTok video player is used. Note that the video will likely fail to
appear when previewing locally, unless you serve the document with something
like \code{servr::httd()}. In limited HTML formats or non-HTML formats, the
TikTok content is rendered into an image with a link in markdown. Use the
helper functions listed below if you want to control the output.
}
\section{Functions}{
\itemize{
\item \code{tiktok_embed}: Embed a TikTok video in R Markdown, or return the oEmbed
API response.

\item \code{tiktok_md}: Format the TikTok video url or \code{tiktok_embed()} result as
markdown

\item \code{tiktok_html}: Format the TikTok video url or \code{tiktok_embed()} result as
HTML
}}

\examples{
tt_url <- "https://www.tiktok.com/@aquickspoonful/video/6890681375431691526"
tt <- tiktok_embed(tt_url)

# in R Markdown, just print the object to embed it
tt

# Or write a plain markdown version
tiktok_md(tt)

# Or as HTML without the full TikTok embedded player shenanigans
tiktok_html(tt, include_player = FALSE)

}
\references{
\url{https://developers.tiktok.com/doc/Embed}
}
